+++ packages/brick_offline_first_with_graphql/CHANGELOG.md
@@ -1,5 +1,8 @@
 ## Unreleased
 
+## 1.1.0
+
+* Use public release versions of `brick_offline_first` and `brick_sqlite`
 * When opening a new `OfflineFirstWithGraphqlRepository#subscribe`, add one event of existing local data to the stream
 * Notify subscribers of an empty payload after deleting